refactor: 자잘한 코드 수정 및 유지보수성 증가를 위한 리팩토링

+++ src/main/java/com/chukchuk/haksa/application/academic/AcademicRecord.java
@@ -99,15 +99,15 @@ public class AcademicRecord {
         return courseEnrollments.getEligibleForRetake();
     }
 
-    public boolean hasPassedCourse(int offeringId) {
+    public boolean hasPassedCourse(Long offeringId) {
         return courseEnrollments.hasPassedCourse(offeringId);
     }
 
-    public boolean isCurrentlyEnrolled(int offeringId) {
+    public boolean isCurrentlyEnrolled(Long offeringId) {
         return courseEnrollments.isCurrentlyEnrolled(offeringId);
     }
 
-    public boolean canTakeCourse(int offeringId) {
+    public boolean canTakeCourse(Long offeringId) {
         return !hasPassedCourse(offeringId) && !isCurrentlyEnrolled(offeringId);
     }
 }
